Agregar la propiedad id en los registros del array tareas
