+++ NEWBIE/TO-DO todo-app-main/js/functionalities/checked.js
