+++ NEWBIE/TO-DO todo-app-main/js/imgs/svg.js
