+++ NEWBIE/TO-DO todo-app-main/js/tareas.js
@@ -1,7 +1,7 @@
 export const tareas = [];
 
-export const insertarDatosTarea = (text) => {
-  const nuevaTarea = { texto: text, finalizado: false };
+export const insertarDatosTarea = (text,id) => {
+  const nuevaTarea = {id:id,texto: text, finalizado: false };
   tareas.push(nuevaTarea);
   console.log(tareas);
 };
@@ -9,3 +9,19 @@ export const obtenerTareas = () => {
   return tareas;
 };
 
+export const cambiarEstadoHomeWork = (tarea, input) => {
+  const tareas = obtenerTareas();
+  if (input.checked) {
+    tarea.finalizado = true;
+  } else {
+    tarea.finalizado = false;
+  }
+
+  tareas.map((t) => {
+    if (t.id === tarea.id) {
+      t.finalizado = tarea.finalizado;
+    }
+
+    return t;
+  });
+};
