+++ NEWBIE/TO-DO todo-app-main/js/utilities.js
@@ -5,3 +5,7 @@ export const createTag = (tagName,className) => {
     };
     return element;
 }
+export const generarIdÚnico = () =>{
+    const idAleatorio = Math.random().toString(8).substring(2,6);
+    return idAleatorio
+  }
